seedgui/requirements.txt
absl-py==1.4.0
amqp==5.1.1
arrow==1.2.3
asgiref==3.5.2
astunparse==1.6.3
async-timeout==4.0.2
billiard==4.1.0
black==23.3.0
cachetools==5.3.0
celery==5.3.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
contourpy==1.0.7
cron-descriptor==1.4.0
cycler==0.11.0
Django==4.1.3
django-admin-interface==0.22.2
django-celery-beat==2.5.0
django-colorfield==0.7.2
django-enum-choices==2.1.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-timezone-field==5.1
flatbuffers==23.1.21
fonttools==4.38.0
gast==0.4.0
gevent==22.10.2
google-auth==2.16.2
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.51.3
h5py==3.8.0
idna==3.4
keras==2.11.0
kiwisolver==1.4.4
kombu==5.3.1
libclang==15.0.6.1
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.0
mypy-extensions==1.0.0
numpy==1.23.4
oauthlib==3.2.2
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging==23.0
pandas==1.5.3
pathspec==0.11.1
Pillow==9.3.0
platformdirs==3.7.0
plotly==5.6.0
proj==0.2.0
prompt-toolkit==3.0.38
protobuf==3.19.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-crontab==2.7.1
python-dateutil==2.8.2
pytz==2022.7.1
PyYAML==6.0
redis==4.5.5
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
sqlparse==0.4.3
tenacity==8.2.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
tomli==2.0.1
tflite-runtime-nightly
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.6
Werkzeug==2.2.3
wrapt==1.15.0
zope.event==4.6
zope.interface==6.0
RPi.GPIO
